pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "meeting-transcriber"
version = "0.1.0"
description = "会議動画から話者識別付き議事録を作成するMCPサーバー & CLIツール"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "whisper", "transcription", "diarization", "meeting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp>=1.0.0",
    "mlx-whisper>=0.4.0",
    "pyannote.audio>=3.1.1",
    "opencv-python>=4.8.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
transcribe = "meeting_transcriber.cli:main"
meeting-transcriber-server = "meeting_transcriber.server:run"

[tool.setuptools.packages.find]
where = ["src"]
